// src/app/entities/producto-interface.ts
export interface ProductoInterface {
  descripcion?: string;
  id_ao?: number;
  id_estado?: number;
  id_ugp: number;
  id_unidad_medida?: number;
  nombre?: string;
  id?: null;
  factor?: number;
  id_proceso?:number;
  indicador?: string;
  seleccionado?:boolean;
  usuario_registro?: string;
  nombre_usuario?: string;
  trama_ae?:string;
  nombre_proceso?:string;
  codigo_proceso?: string;
}